Add a test with pointers and wrappers
This test shows what will ultimately be the goal of this PR: a receiver
is object-safe as long as it can be coerced from a fat pointer to a
thin pointer. The requirement is that if `Self: Unsize<U>`, the
receiver type will implement `CoerceUnsized<FatReceiver>` where
`FatReceiver` is the same as the receiver type, but with `U`
substituted in for `Self`.